Revert "docs: remove superfluous SITE_URL from .env.local"
This reverts commit a9b3086a5edf43091032b20806fee5678977a0bc.

Full stack framework headers require the baseURL to be defined the
same as it is in Convex because the url being http or https affects
whether Better Auth uses secure cookies. A mismatch means the framework
server may end up looking for an insecure cookie when a secure cookie
has been set, or vice versa, leading to a broken implementation.